+++ packages/commonwealth/client/scripts/views/components/Header/UserDropdown/UserDropdownItem.scss
@@ -1,0 +1,37 @@
+@import '../../../../../styles/shared';
+
+.UserDropdownItem {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+
+  .Text {
+    color: $neutral-500;
+  }
+
+  &.isSignedIn {
+    .Text {
+      color: $neutral-700;
+      font-weight: 500;
+    }
+  }
+
+  .identification {
+    overflow: hidden;
+    display: inline-block;
+    text-overflow: ellipsis;
+    max-width: 120px;
+    white-space: nowrap;
+  }
+
+  .check-icon {
+    display: flex;
+
+    svg {
+      fill: $green-600;
+      width: 16px;
+      height: 16px;
+    }
+  }
+}
